+++ test/test/__init__.py
@@ -1,11 +1,16 @@
 # noinspection PyPackageRequirements
+from typing import Optional
+
 import sklearn
 from packaging import version
 
 from test.paths import TEST_CONFIG_YML
 
 
-def check_sklearn_version(minimum: str = "0.21", maximum: str = "0.24"):
+def check_sklearn_version(minimum: Optional[str], maximum: Optional[str] = None):
     """ Utility to check sklearn version against provided string. """
     v_sklearn = version.parse(sklearn.__version__)
-    return version.parse(minimum) <= v_sklearn <= version.parse(maximum)
+    return not (
+        (minimum and version.parse(minimum) > v_sklearn)
+        or (maximum and v_sklearn > version.parse(maximum))
+    )